Add directive#toString() & directive#toDocumentNode()
A recent change to Apollo Server's type definitions caused Typescript to complain about how we were exporting directive type definitions from this library. To solve this, two new methods were added to generate the correct type declarations while also satisfying the type issues with the directive class.

This results in a breaking change and the documentation has been updated to reflect the new requirements for adding directives to a schema.